feat: implement SpiderCacheCoordinator for cache coherence management
feat: add SpiderDependencyAnalyzer for file analysis and dependency resolution

feat: create SpiderGraphCrawler to traverse and build dependency graph

feat: introduce SpiderIndexingCancellation for shared cancellation in indexing

feat: develop SpiderIndexingService for managing reverse index and indexing operations

feat: implement SpiderReferenceLookup for reverse dependency lookup

feat: add SpiderSymbolService for symbol-level analysis and features
